add processes window cleanup
- refactor code
- add unit test
- fix warnings.
