--- nohup.out
@@ -1,8 +1,0 @@
-socktop_agent: TLS enabled. Listening on wss://0.0.0.0:8433/ws
-socktop_agent: TLS enabled. Listening on wss://0.0.0.0:8433/ws
-socktop_agent: TLS enabled. Listening on wss://0.0.0.0:8433/ws
-Error: Address already in use (os error 98)
-socktop_agent: TLS enabled. Listening on wss://0.0.0.0:8433/ws
-Error: Address already in use (os error 98)
-socktop_agent: TLS enabled. Listening on wss://0.0.0.0:8443/ws
-socktop_agent: TLS enabled. Listening on wss://0.0.0.0:8443/ws
