+++ socktop/src/ui/mod.rs
@@ -6,6 +6,10 @@ pub mod gpu;
 pub mod header;
 pub mod mem;
 pub mod modal;
+pub mod modal_connection;
+pub mod modal_format;
+pub mod modal_process;
+pub mod modal_types;
 pub mod net;
 pub mod processes;
 pub mod swap;
